+++ src/main/java/com/imagegrafia/petrolpump/service/PumpService.java
@@ -30,7 +30,7 @@ public class PumpService {
 		return (List<Pump>) pumpRepository.findAll();
 	}
 	
-	public Pump findPumpById(int pumpId) {
+	public Pump findPumpById(Long pumpId) {
 		return pumpRepository.findById(pumpId).get();
 	}
 
@@ -39,7 +39,7 @@ public class PumpService {
 		return pumpRepository.save(pump);
 	}
 
-	public Pump updatePump(int pumpId, Pump pump) {
+	public Pump updatePump(Long pumpId, Pump pump) {
 		pump.setId(pumpId);
 		return pumpRepository.save(pump);
 	}
